added possibility to get file with average temperature

+++ cleanup.sh
@@ -7,7 +7,6 @@ tmpFile=temporary_file.csv
 #decide which file to clenaup
 echo "The file of which city do you want to cleanup?(Please enter just the city name: [City])"
 read file
-echo $file
 
 if [ $file != "Karlstad" ]; then
 	datafile=smhi-opendata_$file.csv
@@ -22,7 +21,7 @@ if [[ ! $file ]]; then
 fi
 
 #remove old file if it exists
-if [[ cleaned_${file}.csv ]]; then
+if [[ -f cleaned_${file}.csv ]]; then
 	echo "Deleting old file csx"
 	rm cleaned_$file.csv
 fi
@@ -40,8 +39,8 @@ awk -F ";" '{print $1";"$2";"$3";"$4}' ${tmpFile}>>cleaned_$file.csv
 #ask if one wants to make a file with only high quality results
 echo "Do you want to make a file with only high Quality results?[y/n]"
 read input
-if input==y; then
-	if [[ cleaned_${file}_highQuality.csv ]]; then
+if [ $input == "y" ]; then
+	if [[ -f cleaned_${file}_highQuality.csv ]]; then
 	echo "Deleting old file"
 	rm cleaned_${file}_highQuality.csv
 	fi
@@ -50,11 +49,41 @@ fgrep "Y" cleaned_${file}.csv | wc -l
 echo "low Quality entries"
 grep "G" cleaned_${file}.csv >> cleaned_${file}_highQuality.csv
 fi
+ 
+#date=`awk -F ";" 'NR==1{print $1}' ${tempFile}`
+#echo $date
 
+#Make a file containing the average tmeperature of a day, Format: date averageTemperature
+Date=1961-01-01
+rm oneDayTemp.txt
+touch oneDay.txt
+touch oneDayTemp.txt
+oneDay=oneDay.txt
+oneDayTemp=oneDayTemp.txt
+#increase starting date by one day for each i
+for i in {0..100}
+do 
+	next_date=$(date +%Y-%m-%d -d "$Date +$i day")
+	#grep only the lines with the same date
+	egrep ^${next_date} ${tmpFile}>${oneDay}
+	#get the number of entries per day
+	lines=`cat $oneDay | wc -l`
+	#make a file with the date and the average of the temperature of thaht day
+	awk -F ";" -v date="$next_date" -v line="$lines" '{sum += $3} END {print date" "sum/line}' ${oneDay}>>${oneDayTemp}
+	rm $oneDay
+	#rm $oneDayTemp
+done
+ 
 # store in easy readable format for c++
-rm easyToRead0600_${file}.txt
-rm easyToRead1200_${file}.txt
-rm easyToRead1800_${file}.txt
+if [[ -f easyToRead0600_${file}.txt ]]; then
+	rm easyToRead0600_${file}.txt
+fi
+if [[ -f easyToRead1200_${file}.txt ]]; then
+	rm easyToRead1200_${file}.txt
+fi
+if [[ -f easyToRead1800_${file}.txt ]]; then
+	rm easyToRead1800_${file}.txt
+fi
 
 #all data at 06:00
 egrep ^[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]\;06 $datafile > ${tmpFile}
@@ -69,5 +98,8 @@ egrep ^[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]\;18 $datafile > ${tmpFile}
 awk -F ";" '{print $1" "$3}' ${tmpFile}>>easyToRead1800_${file}.txt
 
 
-rm $tmpFile
+
+
+
+#rm $tmpFile
